Change expired column type to bigint

+++ database/migrations/0001_01_01_000001_create_cache_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('cache', function (Blueprint $table) {
             $table->string('key')->primary();
             $table->mediumText('value');
-            $table->integer('expiration');
+            $table->bigInteger('expiration');
         });
 
         Schema::create('cache_locks', function (Blueprint $table) {
